Revert "Implement IMaskinportenClient interface (#669)" (#713)
This reverts commit d6bfcccba026c40f62e7b110d2501fbb55910053.
