+++ src/Altinn.App.Api/Extensions/ServiceCollectionExtensions.cs
@@ -7,8 +7,6 @@ using Altinn.App.Api.Infrastructure.Telemetry;
 using Altinn.App.Core.Constants;
 using Altinn.App.Core.Extensions;
 using Altinn.App.Core.Features;
-using Altinn.App.Core.Features.Maskinporten;
-using Altinn.App.Core.Features.Maskinporten.Models;
 using Altinn.Common.PEP.Authorization;
 using Altinn.Common.PEP.Clients;
 using AltinnCore.Authentication.JwtCookie;
@@ -17,7 +15,6 @@ using Microsoft.ApplicationInsights.Extensibility;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Microsoft.Extensions.Options;
 using Microsoft.FeatureManagement;
 using Microsoft.IdentityModel.Tokens;
 using OpenTelemetry;
@@ -68,7 +65,6 @@ public static class ServiceCollectionExtensions
 
         services.AddPlatformServices(config, env);
         services.AddAppServices(config, env);
-        services.AddMaskinportenClient();
         services.ConfigureDataProtection();
 
         var useOpenTelemetrySetting = config.GetValue<bool?>("AppSettings:UseOpenTelemetry");
@@ -98,50 +94,6 @@ public static class ServiceCollectionExtensions
         services.AddHttpClient<AuthorizationApiClient>();
 
         services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
-    }
-
-    /// <summary>
-    /// <para>
-    /// Configures the <see cref="MaskinportenClient"/> service with a configuration object which will be static for the lifetime of the service.
-    /// </para>
-    /// <para>
-    /// If you have already provided a <see cref="MaskinportenSettings"/> configuration, either manually or
-    /// implicitly via <see cref="WebHostBuilderExtensions.ConfigureAppWebHost"/>, this will be overridden.
-    /// </para>
-    /// </summary>
-    /// <param name="services">The service collection</param>
-    /// <param name="configureOptions">
-    /// Action delegate that provides <see cref="MaskinportenSettings"/> configuration for the <see cref="MaskinportenClient"/> service
-    /// </param>
-    public static IServiceCollection ConfigureMaskinportenClient(
-        this IServiceCollection services,
-        Action<MaskinportenSettings> configureOptions
-    )
-    {
-        services.AddOptions<MaskinportenSettings>().Configure(configureOptions).ValidateDataAnnotations();
-
-        return services;
-    }
-
-    /// <summary>
-    /// Adds a singleton <see cref="AddMaskinportenClient"/> service to the service collection.
-    /// Binds <see cref="MaskinportenSettings"/>, either from `appsettings.json` or `maskinporten-settings.json` (if found).
-    /// <br/><br/>Note: This binding happens in <see cref="WebHostBuilderExtensions.AddMaskinportenSettingsFile"/>.
-    /// </summary>
-    /// <param name="services">The service collection</param>
-    private static IServiceCollection AddMaskinportenClient(this IServiceCollection services)
-    {
-        if (services.GetOptionsDescriptor<MaskinportenSettings>() is null)
-        {
-            services
-                .AddOptions<MaskinportenSettings>()
-                .BindConfiguration("MaskinportenSettings")
-                .ValidateDataAnnotations();
-        }
-
-        services.AddSingleton<IMaskinportenClient, MaskinportenClient>();
-
-        return services;
     }
 
     /// <summary>
@@ -458,19 +410,6 @@ public static class ServiceCollectionExtensions
         });
 
         services.TryAddSingleton<ValidateAntiforgeryTokenIfAuthCookieAuthorizationFilter>();
-    }
-
-    private static IServiceCollection RemoveOptions<TOptions>(this IServiceCollection services)
-        where TOptions : class
-    {
-        var descriptor = services.GetOptionsDescriptor<TOptions>();
-
-        if (descriptor is not null)
-        {
-            services.Remove(descriptor);
-        }
-
-        return services;
     }
 
     private static (string? Key, string? ConnectionString) GetAppInsightsConfig(
