+++ src/Altinn.App.Api/Extensions/WebHostBuilderExtensions.cs
@@ -1,7 +1,4 @@
 using Altinn.App.Core.Extensions;
-using Altinn.App.Core.Features.Maskinporten;
-using Altinn.App.Core.Features.Maskinporten.Models;
-using Microsoft.Extensions.FileProviders;
 
 namespace Altinn.App.Api.Extensions;
 
@@ -30,37 +27,8 @@ public static class WebHostBuilderExtensions
                 }
 
                 configBuilder.AddInMemoryCollection(config);
-
-                configBuilder.AddMaskinportenSettingsFile(context);
-
                 configBuilder.LoadAppConfig(args);
             }
         );
-    }
-
-    private static IConfigurationBuilder AddMaskinportenSettingsFile(
-        this IConfigurationBuilder configurationBuilder,
-        WebHostBuilderContext context
-    )
-    {
-        string jsonProvidedPath =
-            context.Configuration.GetValue<string>("MaskinportenSettingsFilepath")
-            ?? "/mnt/app-secrets/maskinporten-settings.json";
-        string jsonAbsolutePath = Path.GetFullPath(jsonProvidedPath);
-
-        if (File.Exists(jsonAbsolutePath))
-        {
-            string jsonDir = Path.GetDirectoryName(jsonAbsolutePath) ?? string.Empty;
-            string jsonFile = Path.GetFileName(jsonAbsolutePath);
-
-            configurationBuilder.AddJsonFile(
-                provider: new PhysicalFileProvider(jsonDir),
-                path: jsonFile,
-                optional: true,
-                reloadOnChange: true
-            );
-        }
-
-        return configurationBuilder;
     }
 }
