+++ src/Altinn.App.Core/EFormidling/Implementation/EformidlingStatusCheckEventHandler2.cs
@@ -25,11 +25,7 @@ public class EformidlingStatusCheckEventHandler2 : IEventHandler
     private readonly IEFormidlingClient _eFormidlingClient;
     private readonly ILogger<EformidlingStatusCheckEventHandler2> _logger;
     private readonly IHttpClientFactory _httpClientFactory;
-
-#pragma warning disable CS0618 // 'member' is obsolete:
     private readonly IMaskinportenTokenProvider _maskinportenTokenProvider;
-#pragma warning restore CS0618
-
     private readonly PlatformSettings _platformSettings;
     private readonly GeneralSettings _generalSettings;
 
@@ -40,9 +36,7 @@ public class EformidlingStatusCheckEventHandler2 : IEventHandler
         IEFormidlingClient eFormidlingClient,
         IHttpClientFactory httpClientFactory,
         ILogger<EformidlingStatusCheckEventHandler2> logger,
-#pragma warning disable CS0618 // 'member' is obsolete:
         IMaskinportenTokenProvider maskinportenTokenProvider,
-#pragma warning restore CS0618
         IOptions<PlatformSettings> platformSettings,
         IOptions<GeneralSettings> generalSettings
     )
