+++ src/Altinn.App.Core/Extensions/ServiceCollectionExtensions.cs
@@ -58,7 +58,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Options;
 using Newtonsoft.Json.Linq;
 using IProcessEngine = Altinn.App.Core.Internal.Process.IProcessEngine;
 using IProcessReader = Altinn.App.Core.Internal.Process.IProcessReader;
@@ -106,7 +105,6 @@ public static class ServiceCollectionExtensions
 #pragma warning restore CS0618 // Type or member is obsolete
         services.AddHttpClient<IProcessClient, ProcessClient>();
         services.AddHttpClient<IPersonClient, PersonClient>();
-        services.AddHybridCache();
 
         services.TryAddTransient<IUserTokenProvider, UserTokenProvider>();
         services.TryAddTransient<IAccessTokenGenerator, AccessTokenGenerator>();
@@ -368,20 +366,5 @@ public static class ServiceCollectionExtensions
     {
         services.TryAddTransient<IFileValidationService, FileValidationService>();
         services.TryAddTransient<IFileValidatorFactory, FileValidatorFactory>();
-    }
-
-    internal static IEnumerable<ServiceDescriptor> GetOptionsDescriptors<TOptions>(this IServiceCollection services)
-        where TOptions : class
-    {
-        return services.Where(d =>
-            d.ServiceType == typeof(IConfigureOptions<TOptions>)
-            || d.ServiceType == typeof(IOptionsChangeTokenSource<TOptions>)
-        );
-    }
-
-    internal static ServiceDescriptor? GetOptionsDescriptor<TOptions>(this IServiceCollection services)
-        where TOptions : class
-    {
-        return services.GetOptionsDescriptors<TOptions>().FirstOrDefault();
     }
 }
