--- src/Altinn.App.Core/Features/Maskinporten/Models/TokenCacheEntry.cs
@@ -1,6 +1,0 @@
-using System.ComponentModel;
-
-namespace Altinn.App.Core.Features.Maskinporten.Models;
-
-[ImmutableObject(true)]
-internal sealed record TokenCacheEntry(MaskinportenTokenResponse Token, TimeSpan Expiration, bool HasSetExpiration);
