+++ src/Altinn.App.Core/Features/Telemetry.cs
@@ -75,7 +75,6 @@ public sealed partial class Telemetry : IDisposable
             InitNotifications(context);
             InitProcesses(context);
             InitValidation(context);
-            InitMaskinporten(context);
 
             // NOTE: This Telemetry class is registered as a singleton
             // Metrics could be kept in fields of the respective objects that use them for instrumentation
