+++ test/Altinn.App.Api.Tests/Telemetry/TelemetryConfigurationTests.cs
@@ -1,6 +1,5 @@
 using System.Diagnostics.Tracing;
 using System.Reflection;
-using Altinn.App.Api.Tests.TestUtils;
 using Altinn.App.Core.Features;
 using Microsoft.ApplicationInsights;
 using Microsoft.ApplicationInsights.Channel;
@@ -151,7 +150,7 @@ public class TelemetryConfigurationTests
             )
             .Build();
 
-        Altinn.App.Api.Extensions.ServiceCollectionExtensions.AddAltinnAppServices(services, config, env);
+        Extensions.ServiceCollectionExtensions.AddAltinnAppServices(services, config, env);
         services.AddApplicationInsightsTelemetryProcessor<TelemetryProcessor>();
 
         await using (var sp = services.BuildServiceProvider())
@@ -206,7 +205,7 @@ public class TelemetryConfigurationTests
         // Hopefully we can remove all this soon
         services.AddKeyedSingleton<ITelemetryProcessor, TelemetryProcessor>("test");
 
-        Altinn.App.Api.Extensions.ServiceCollectionExtensions.AddAltinnAppServices(services, config, env);
+        Extensions.ServiceCollectionExtensions.AddAltinnAppServices(services, config, env);
 
         await using (var sp = services.BuildServiceProvider())
         {
@@ -221,6 +220,36 @@ public class TelemetryConfigurationTests
         Assert.NotEmpty(events.Where(e => errorLevels.Contains(e.Level)));
     }
 
+    private WebApplication BuildApp(
+        IEnumerable<KeyValuePair<string, string?>> configData,
+        Action<IServiceCollection>? registerCustomAppServices = null
+    )
+    {
+        // Here we follow the order of operations currently present in the Program.cs generated by the template for apps,
+        // which is the following:
+
+        // 0. WebApplication.CreateBuilder()
+        WebApplicationBuilder builder = WebApplication.CreateBuilder();
+        builder.Environment.EnvironmentName = "Development";
+        builder.Configuration.AddInMemoryCollection(configData);
+        // 1. AddAltinnAppControllersWithViews
+        Extensions.ServiceCollectionExtensions.AddAltinnAppControllersWithViews(builder.Services);
+        // 2. RegisterCustomAppServices
+        registerCustomAppServices?.Invoke(builder.Services);
+        // 3. AddAltinnAppServices
+        Extensions.ServiceCollectionExtensions.AddAltinnAppServices(
+            builder.Services,
+            builder.Configuration,
+            builder.Environment
+        );
+        // 4. ConfigureAppWebHost
+        Extensions.WebHostBuilderExtensions.ConfigureAppWebHost(builder.WebHost, []);
+        // 5. UseAltinnAppCommonConfiguration
+        var app = builder.Build();
+        Extensions.WebApplicationBuilderExtensions.UseAltinnAppCommonConfiguration(app);
+        return app;
+    }
+
     [Fact]
     public async Task OpenTelemetry_Registers_Correctly_When_Enabled()
     {
@@ -230,7 +259,7 @@ public class TelemetryConfigurationTests
             new("AppSettings:UseOpenTelemetry", "true"),
         ];
         Telemetry? telemetry = null;
-        await using (var app = AppBuilder.Build(configData: configData))
+        await using (var app = BuildApp(configData))
         {
             var telemetryClient = app.Services.GetService<TelemetryClient>();
             Assert.Null(telemetryClient);
@@ -247,7 +276,7 @@ public class TelemetryConfigurationTests
     public async Task OpenTelemetry_Does_Not_Register_By_Default()
     {
         List<KeyValuePair<string, string?>> configData = [new("ApplicationInsights:InstrumentationKey", "test"),];
-        await using (var app = AppBuilder.Build(configData: configData))
+        await using (var app = BuildApp(configData))
         {
             var telemetryClient = app.Services.GetService<TelemetryClient>();
             Assert.NotNull(telemetryClient);
@@ -265,7 +294,7 @@ public class TelemetryConfigurationTests
             new("ApplicationInsights:InstrumentationKey", "test"),
             new("AppSettings:UseOpenTelemetry", "true"),
         ];
-        await using var app = AppBuilder.Build(configData: configData);
+        await using var app = BuildApp(configData);
 
         var traceProvider = app.Services.GetRequiredService<TracerProvider>();
 
@@ -281,7 +310,7 @@ public class TelemetryConfigurationTests
             new("ApplicationInsights:InstrumentationKey", "test"),
             new("AppSettings:UseOpenTelemetry", "true"),
         ];
-        await using var app = AppBuilder.Build(configData: configData);
+        await using var app = BuildApp(configData);
 
         var options = app.Services.GetRequiredService<IOptions<PeriodicExportingMetricReaderOptions>>().Value;
 
@@ -298,8 +327,8 @@ public class TelemetryConfigurationTests
             new("AppSettings:UseOpenTelemetry", "true"),
         ];
         var samplerToUse = new ParentBasedSampler(new AlwaysOnSampler());
-        await using var app = AppBuilder.Build(
-            configData: configData,
+        await using var app = BuildApp(
+            configData,
             registerCustomAppServices: services =>
             {
                 services.ConfigureOpenTelemetryTracerProvider(builder =>
@@ -326,8 +355,8 @@ public class TelemetryConfigurationTests
 
         var intervalToUse = 5_000;
         var timeoutToUse = 4_000;
-        await using var app = AppBuilder.Build(
-            configData: configData,
+        await using var app = BuildApp(
+            configData,
             registerCustomAppServices: services =>
             {
                 services.Configure<PeriodicExportingMetricReaderOptions>(options =>
