--- test/Altinn.App.Api.Tests/TestUtils/AppBuilder.cs
@@ -1,50 +1,0 @@
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-
-namespace Altinn.App.Api.Tests.TestUtils;
-
-public static class AppBuilder
-{
-    public static WebApplication Build(
-        WebApplicationBuilder? builder = default,
-        IEnumerable<KeyValuePair<string, string?>>? configData = default,
-        Action<IServiceCollection>? registerCustomAppServices = default
-    )
-    {
-        // Here we follow the order of operations currently present in the Program.cs generated by the template for apps,
-        // which is the following:
-
-        // 0. WebApplication.CreateBuilder()
-        builder ??= WebApplication.CreateBuilder();
-        builder.Environment.EnvironmentName = "Development";
-
-        if (configData is not null)
-        {
-            builder.Configuration.AddInMemoryCollection(configData);
-        }
-
-        // 1. AddAltinnAppControllersWithViews
-        Altinn.App.Api.Extensions.ServiceCollectionExtensions.AddAltinnAppControllersWithViews(builder.Services);
-
-        // 2. RegisterCustomAppServices
-        registerCustomAppServices?.Invoke(builder.Services);
-
-        // 3. AddAltinnAppServices
-        Altinn.App.Api.Extensions.ServiceCollectionExtensions.AddAltinnAppServices(
-            builder.Services,
-            builder.Configuration,
-            builder.Environment
-        );
-
-        // 4. ConfigureAppWebHost
-        Altinn.App.Api.Extensions.WebHostBuilderExtensions.ConfigureAppWebHost(builder.WebHost, []);
-
-        // 5. UseAltinnAppCommonConfiguration
-        var app = builder.Build();
-        Altinn.App.Api.Extensions.WebApplicationBuilderExtensions.UseAltinnAppCommonConfiguration(app);
-
-        return app;
-    }
-}
